requirements.txt
# AWS Bedrock (All models via Bedrock)
boto3>=1.34.0
botocore>=1.34.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Async
asyncio>=3.4.3
aiohttp>=3.9.0

# LangGraph & LangChain
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0

# Data validation
pydantic>=2.5.0

# Web framework
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0

# Frontend
streamlit>=1.30.0

# Vector Database (optional for MVP)
# pinecone-client>=3.0.0

# Utilities
python-dotenv>=1.0.0
python-json-logger>=2.0.7
